Only call teardown once per outer block
The deploy method can be called within the a deploy block. When this happens, `teardown!` triggers an immediate app deletion which will cause other code within that block to be deleted.

If I could re-write all of this, I would use a different abstraction instead of having App be the god-object. Since I'm living with this existing domain model, we can record the outermost block ID and only fire teardown when that block is exiting. This method ensures that teardown is only ever called once.
