+++ lib/hatchet/version.rb
@@ -1,3 +1,3 @@
 module Hatchet
-  VERSION = "8.0.1"
+  VERSION = "8.0.2"
 end
